requirements.txt
Flask
Flask-PyMongo
flask-cors
pymongo[srv]
python-dotenv
numpy
opencv-python-headless
scikit-learn
gunicorn
scikit-image